requirements.txt
Flask~=2.2
pyodbc~=4.0
python-dotenv